+++ cli/bpmetadata/cmd.go
@@ -12,6 +12,7 @@ import (
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 	"google.golang.org/protobuf/encoding/protojson"
+	"google.golang.org/protobuf/proto"
 	"sigs.k8s.io/yaml"
 )
 
@@ -127,7 +128,7 @@ func generate(cmd *cobra.Command, args []string) error {
 	}
 
 	if len(errors) > 0 {
-		return fmt.Errorf(strings.Join(errors, "\n"))
+		return fmt.Errorf("%s", strings.Join(errors, "\n"))
 	}
 
 	Log.Info("metadata generated successfully")
@@ -229,8 +230,11 @@ func CreateBlueprintMetadata(bpPath string, bpMetadataObj *BlueprintMetadata) (*
 		return nil, fmt.Errorf("error creating blueprint info: %w", err)
 	}
 
+	var existingInterfaces *BlueprintInterface
 	if bpMetadataObj.Spec.Interfaces == nil {
 		bpMetadataObj.Spec.Interfaces = &BlueprintInterface{}
+	} else {
+		existingInterfaces = proto.Clone(bpMetadataObj.Spec.Interfaces).(*BlueprintInterface)
 	}
 
 	// create blueprint interfaces i.e. variables & outputs
@@ -238,6 +242,9 @@ func CreateBlueprintMetadata(bpPath string, bpMetadataObj *BlueprintMetadata) (*
 	if err != nil {
 		return nil, fmt.Errorf("error creating blueprint interfaces: %w", err)
 	}
+
+    // Merge existing connections (if any) into the newly generated interfaces
+    mergeExistingConnections(bpMetadataObj.Spec.Interfaces, existingInterfaces)
 
 	// get blueprint requirements
 	rolesCfgPath := path.Join(repoDetails.Source.BlueprintRootPath, tfRolesFileName)