+++ tflint-ruleset-blueprint/rules/terraform_required_version_range.go
@@ -12,55 +12,55 @@ import (
 	"github.com/terraform-linters/tflint-plugin-sdk/tflint"
 )
 
-// TerraformRequiredVersion checks if a module has a terraform required_version within valid range.
-type TerraformRequiredVersion struct {
+// TerraformRequiredVersionRange checks if a module has a terraform required_version within valid range.
+type TerraformRequiredVersionRange struct {
 	tflint.DefaultRule
 }
 
-// TerraformRequiredVersionConfig is a config of TerraformRequiredVersion
-type TerraformRequiredVersionConfig struct {
+// TerraformRequiredVersionRangeConfig is a config of TerraformRequiredVersionRange
+type TerraformRequiredVersionRangeConfig struct {
 	MinVersion string `hclext:"min_version,optional"`
 	MaxVersion string `hclext:"max_version,optional"`
 }
 
-// NewTerraformRequiredVersion returns a new rule.
-func NewTerraformRequiredVersion() *TerraformRequiredVersion {
-	return &TerraformRequiredVersion{}
+// NewTerraformRequiredVersionRange returns a new rule.
+func NewTerraformRequiredVersionRange() *TerraformRequiredVersionRange {
+	return &TerraformRequiredVersionRange{}
 }
 
 // Name returns the rule name.
-func (r *TerraformRequiredVersion) Name() string {
-	return "terraform_required_version"
+func (r *TerraformRequiredVersionRange) Name() string {
+	return "terraform_required_version_range"
 }
 
 // Enabled returns whether the rule is enabled by default.
-func (r *TerraformRequiredVersion) Enabled() bool {
+func (r *TerraformRequiredVersionRange) Enabled() bool {
 	return false
 }
 
 // Severity returns the rule severity.
-func (r *TerraformRequiredVersion) Severity() tflint.Severity {
+func (r *TerraformRequiredVersionRange) Severity() tflint.Severity {
 	return tflint.ERROR
 }
 
 // Link returns the rule reference link
-func (r *TerraformRequiredVersion) Link() string {
+func (r *TerraformRequiredVersionRange) Link() string {
 	return "https://googlecloudplatform.github.io/samples-style-guide/#language-specific"
 }
 
 const (
-	minimumTerraformRequiredVersion = "1.3"
-	maximumTerraformRequiredVersion = "1.5"
+	minimumTerraformRequiredVersionRange = "1.3"
+	maximumTerraformRequiredVersionRange = "1.5"
 )
 
 // Checks if a module has a terraform required_version within valid range.
-func (r *TerraformRequiredVersion) Check(runner tflint.Runner) error {
-	config := &TerraformRequiredVersionConfig{}
+func (r *TerraformRequiredVersionRange) Check(runner tflint.Runner) error {
+	config := &TerraformRequiredVersionRangeConfig{}
 	if err := runner.DecodeRuleConfig(r.Name(), config); err != nil {
 		return err
 	}
 
-	minVersion := minimumTerraformRequiredVersion
+	minVersion := minimumTerraformRequiredVersionRange
 	if config.MinVersion != "" {
 		if _, err := version.NewSemver(config.MinVersion); err != nil {
 			return err
@@ -68,7 +68,7 @@ func (r *TerraformRequiredVersion) Check(runner tflint.Runner) error {
 		minVersion = config.MinVersion
 	}
 
-	maxVersion := maximumTerraformRequiredVersion
+	maxVersion := maximumTerraformRequiredVersionRange
 	if config.MaxVersion != "" {
 		if _, err := version.NewSemver(config.MaxVersion); err != nil {
 			return err
@@ -145,10 +145,16 @@ func (r *TerraformRequiredVersion) Check(runner tflint.Runner) error {
 	}
 
 	for _, block := range content.Blocks {
+		requiredVersion, exists := block.Body.Attributes["required_version"]
+		if !exists {
+			logger.Info(fmt.Sprintf("terraform block does not contain required_version: %s", block.DefRange))
+			continue
+		}
+
 		var raw_terraform_required_version string
-		diags := gohcl.DecodeExpression(block.Body.Attributes["required_version"].Expr, nil, &raw_terraform_required_version)
+		diags := gohcl.DecodeExpression(requiredVersion.Expr, nil, &raw_terraform_required_version)
 		if diags.HasErrors() {
-			return fmt.Errorf("failed to decode terraform required_version %q: %v", block.Labels[0], diags.Error())
+			return fmt.Errorf("failed to decode terraform block required_version: %v", diags.Error())
 		}
 
 		constraints, err := version.NewConstraint(raw_terraform_required_version)
@@ -156,7 +162,6 @@ func (r *TerraformRequiredVersion) Check(runner tflint.Runner) error {
 			return err
 		}
 
-		//TODO: add option for repository exemptions
 		if !((constraints.Check(minimum_required_version) || constraints.Check(maximum_required_version)) && !constraints.Check(below_required_version)) {
 			//TODO: use EmitIssueWithFix()
 			err := runner.EmitIssue(r, fmt.Sprintf("required_version is not inclusive of the the minimum %q and maximum %q terraform required_version: %q", minVersion, maxVersion, constraints.String()), block.DefRange)
@@ -164,7 +169,6 @@ func (r *TerraformRequiredVersion) Check(runner tflint.Runner) error {
 				return err
 			}
 		}
-
 	}
 
 	return nil